feat: update theme chooser (#3587)
